[AZURE_AD_AUTH] Implementation of Ad based authentication for azure storage. (#1679)
This PR adds the functionality of AD based authentication for Blob Storage in ambry-cloud. The current authentication method using connection string is unsafe because connection string is visible to anyone who is able to login the account. With AD based authentication, permissions can be assigned to make sure that the client secret is not visible to everyone.

To add this functionality, a class StorageClientFactory provides the base implementation for Azure BlobServiceClient. The client is built in one of ConnectionStringBasedStorageClientFactory or ADAuthBasedStorageClientFactory classes based on config.